refactor(workflows): clarify extension ownership
