+++ .changeset/patch-ms3it38r-32745d.md
@@ -3,4 +3,4 @@
 "@howaboua/pi-subagent-review": patch
 ---
 
-Use the configurable lightweight summary model for tree summaries.
+Use the configurable lightweight summary model for tree summaries. Clarify workflow, review-context, summary-session, and RPC protocol ownership.
