+++ package.json
@@ -16,8 +16,8 @@
     "version-packages": "changeset version && bun run changelog:polish && bun run aggregate:sync && bun run changelog:sync",
     "release": "changeset publish",
     "check:changed": "bun run scripts/check-changed.mjs",
-    "knip": "knip --workspace @howaboua/pi-codex-conversion-lite --workspace @howaboua/pi-auto-trees --no-progress --include files,dependencies,unlisted,unresolved,cycles",
-    "knip:exports": "knip --workspace @howaboua/pi-codex-conversion-lite --workspace @howaboua/pi-auto-trees --no-progress --exports",
+    "knip": "knip --workspace @howaboua/pi-codex-conversion-lite --workspace @howaboua/pi-auto-trees --workspace @howaboua/pi-subagent-review --no-progress --include files,dependencies,unlisted,unresolved,cycles",
+    "knip:exports": "knip --workspace @howaboua/pi-codex-conversion-lite --workspace @howaboua/pi-auto-trees --workspace @howaboua/pi-subagent-review --no-progress --exports",
     "changeset:aggregates": "bun run scripts/add-aggregate-changesets.mjs",
     "hooks:install": "git config core.hooksPath .githooks",
     "changeset:check": "bun run scripts/require-changeset.mjs origin/main",
